requirements.txt
# Web Framework
flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0

# Environment & Config
python-dotenv==1.0.0
pyyaml==6.0.1

# Core TUI
textual==0.47.1
rich==13.7.0

# Audio capture
sounddevice==0.4.6
numpy==1.26.3
scipy==1.11.4
webrtcvad==2.0.10

# Speech-to-Text
faster-whisper==0.10.0
vosk==0.3.45

# LLM
openai==1.12.0
httpx==0.26.0

# Text processing & search
rapidfuzz==3.6.1
sentence-transformers==2.3.1

# Export
python-docx==1.1.0

# Utilities
python-dateutil==2.8.2
